reorder sections and warnings
